intro e variables

+++ exercises/01_variables/variables6.rs
@@ -1,5 +1,6 @@
 // TODO: Change the line below to fix the compiler error.
-const NUMBER : i32 = 3;
+// DONE
+const NUMBER : i64 = 3;
 
 fn main() {
     println!("Number: {NUMBER}");
